requirements.txt
torch
torchaudio
fastapi
uvicorn
python-multipart
git+https://github.com/m-bain/whisperx.git
python-dotenv
